delete image when modify image only

+++ controllers/sauces.js
@@ -32,21 +32,24 @@ exports.createThing = (req, res, next) => {
 
 /* -- modifie une sauce et l'image si nécessaire-- */
 exports.modifyThing = (req, res, next) => {
-  const sauceObject = req.file
-    ? {
-      ...JSON.parse(req.body.sauce),
-      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-    } : { ...req.body }
-  Sauce.findOne({ _id: req.params.id })
-    .then(sauce => {
-      const filename = sauce.imageUrl.split('/images/')[1]
-      fs.unlink(`images/${filename}`, () => {
-        Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, $set: { likes: 0, dislikes: 0, usersLiked: [], usersDisliked: [] }, _id: req.params.id })
-          .then(() => res.status(200).json({ message: ' Sauce modifiée' }))
-          .catch(error => res.status(400).json({ error }))
+  if (req.file) {
+    JSON.parse(req.body.sauce)
+    const imageUrl = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    Sauce.findOne({ _id: req.params.id })
+      .then(sauce => {
+        const filename = sauce.imageUrl.split('/images/')[1]
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.updateOne({ _id: req.params.id }, { ...req.file, $set: { imageUrl: imageUrl, likes: 0, dislikes: 0, usersLiked: [], usersDisliked: [] }, _id: req.params.id })
+            .then(() => res.status(200).json({ message: ' Sauce modifiée' }))
+            .catch(error => res.status(400).json({ error }))
+        })
       })
-    })
-    .catch(error => res.status(500).json({ error }))
+      .catch(error => res.status(500).json({ error }))
+  } else {
+    Sauce.updateOne({ _id: req.params.id }, { ...req.file, $set: { likes: 0, dislikes: 0, usersLiked: [], usersDisliked: [] }, _id: req.params.id })
+      .then(() => res.status(200).json({ message: ' Sauce modifiée' }))
+      .catch(error => res.status(400).json({ error }))
+  }
 }
 
 /* -- supprime une sauce et son image du dossier images -- */
